Implemented Neural Network
I just implemented the MLPClassifier to study the relationship between the goal and several other metadata. Not all columns were involved, you may see the column numbers to see which metadata did I use, or, if needed, I can provide you a list of metadata. 

I only used the 1 trunk in this project, which contains 1000 lines. Since there's no need to increase the number of rows since the prediction probabilities and accuracies under some variables have already achieved to 99 percent.

ChatGPT was involved in the "Figure" part, and helped me to revise errors that happened in the training section.